# docs/swagger.json
{
    "schemes": [
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "API documentation for Gloo",
        "title": "Gloo API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "email": "support@gloo.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "gloo-api-production.up.railway.app",
    "basePath": "/api",
    "paths": {
...
        "/recipes": {
            "get": {
                "description": "Get all recipes by search and category",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Recipes"
                ],
                "summary": "Get all recipes",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Search term for recipes",
                        "name": "searchParam",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Category filter for recipes",
                        "name": "category",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully retrieved recipes",
                        "schema": {
                            "$ref": "#/definitions/models.GetAllRecipesResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorBadRequestResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorNotFoundResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorInternalServerErrorResponse"
                        }
                    }
                }
            }
        },
        "/recipes/admin/{organizationId}": {
            "get": {
                "description": "Get a specific recipe by its ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Recipes"
                ],
                "summary": "Get recipe",
                "parameters": [
                    {
                        "minimum": 1,
                        "type": "integer",
                        "description": "Organization ID",
                        "name": "organizationId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully retrieved recipe",
                        "schema": {
                            "$ref": "#/definitions/models.GetAllRecipeAdminResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorBadRequestResponse"
                        }
                    },
                    "404": {
                        "description": "Recipe not found",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorNotFoundResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorInternalServerErrorResponse"
                        }
                    }
                }
            }
        },
        "/recipes/admin/{recipeId}/{adminId}": {
            "post": {
                "description": "Update a specific recipe by its ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Recipes"
                ],
                "summary": "Update recipe",
                "parameters": [
                    {
                        "minimum": 1,
                        "type": "integer",
                        "description": "Recipe ID",
                        "name": "recipeId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "minimum": 1,
                        "type": "integer",
                        "description": "Admin ID",
                        "name": "adminId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully updated recipe",
                        "schema": {
                            "$ref": "#/definitions/models.UpdateResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorBadRequestResponse"
                        }
                    },
                    "404": {
                        "description": "Recipe not found",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorNotFoundResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorInternalServerErrorResponse"
                        }
                    }
                }
            }
        },
        "/recipes/{id}": {
            "get": {
                "description": "Get a specific recipe by its ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Recipes"
                ],
                "summary": "Get recipe by ID",
                "parameters": [
                    {
                        "minimum": 1,
                        "type": "integer",
                        "description": "Recipe ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully retrieved recipe",
                        "schema": {
                            "$ref": "#/definitions/models.GetRecipeByIdResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorBadRequestResponse"
                        }
                    },
                    "404": {
                        "description": "Recipe not found",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorNotFoundResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorInternalServerErrorResponse"
                        }
                    }
                }
            },
            "put": {
                "description": "Update a specific recipe by its ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Recipes"
                ],
                "summary": "Update recipe",
                "parameters": [
                    {
                        "minimum": 1,
                        "type": "integer",
                        "description": "Recipe ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Recipe",
                        "name": "recipe",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.UpdateRecipeRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully updated recipe",
                        "schema": {
                            "$ref": "#/definitions/models.UpdateRecipeRequest"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorBadRequestResponse"
                        }
                    },
                    "404": {
                        "description": "Recipe not found",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorNotFoundResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorInternalServerErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete a specific recipe by its ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Recipes"
                ],
                "summary": "Delete recipe",
                "parameters": [
                    {
                        "minimum": 1,
                        "type": "integer",
                        "description": "Recipe ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully deleted recipe",
                        "schema": {
                            "$ref": "#/definitions/models.DeleteRecipeRequest"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorBadRequestResponse"
                        }
                    },
                    "404": {
                        "description": "Recipe not found",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorNotFoundResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorInternalServerErrorResponse"
                        }
                    }
                }
            }
        },
        "/recipes/{userId}": {
            "post": {
                "description": "Create a new recipe",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Recipes"
                ],
                "summary": "Create recipe",
                "parameters": [
                    {
                        "minimum": 1,
                        "type": "integer",
                        "description": "User ID",
                        "name": "userId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Recipe",
                        "name": "recipe",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.CreateRecipeRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully created recipe",
                        "schema": {
                            "$ref": "#/definitions/models.CreateRecipeRequest"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorBadRequestResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorInternalServerErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "models.ErrorBadRequestResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "Bad Request"
                },
                "status": {
                    "type": "string",
                    "example": "error"
                },
                "status_code": {
                    "type": "integer",
                    "example": 400
                },
                "success": {
                    "type": "boolean",
                    "example": false
                }
            }
        },
        "models.ErrorInternalServerErrorResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "Internal Server Error"
                },
                "status": {
                    "type": "string",
                    "example": "error"
                },
                "status_code": {
                    "type": "integer",
                    "example": 500
                },
                "success": {
                    "type": "boolean",
                    "example": false
                }
            }
        },
        "models.ErrorNotFoundResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "Not Found"
                },
                "status": {
                    "type": "string",
                    "example": "error"
                },
                "status_code": {
                    "type": "integer",
                    "example": 404
                },
                "success": {
                    "type": "boolean",
                    "example": false
                }
            }
        },
...
        "models.GetAllRecipeAdminResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "type": "object",
                        "properties": {
                            "description": {
                                "type": "string",
                                "example": "A feast for the senses"
                            },
                            "estimated_time": {
                                "type": "string",
                                "example": "20 minutes"
                            },
                            "id": {
                                "type": "integer",
                                "example": 1
                            },
                            "image_url": {
                                "type": "string",
                                "example": "https://example.com/image.png"
                            },
                            "title": {
                                "type": "string",
                                "example": "Chicken Wings"
                            }
                        }
                    }
                },
                "message": {
                    "type": "string",
                    "example": "Successfully retrieved recipe"
                },
                "status": {
                    "type": "string",
                    "example": "success"
                },
                "status_code": {
                    "type": "integer",
                    "example": 200
                },
                "success": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
        "models.GetAllRecipesResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "type": "object",
                        "properties": {
                            "category": {
                                "type": "string",
                                "example": "Chicken"
                            },
                            "created_by": {
                                "type": "string",
                                "example": "@facu.potti"
                            },
                            "description": {
                                "type": "string",
                                "example": "A feast for the senses"
                            },
                            "estimated_time": {
                                "type": "string",
                                "example": "20 minutes"
                            },
                            "id": {
                                "type": "integer",
                                "example": 1
                            },
                            "image_url": {
                                "type": "string",
                                "example": "https://example.com/image.png"
                            },
                            "searchParam": {
                                "type": "string",
                                "example": "chicken"
                            },
                            "title": {
                                "type": "string",
                                "example": "Chicken Wings"
                            },
                            "total_comments": {
                                "type": "integer",
                                "example": 64
                            },
                            "total_favorites": {
                                "type": "integer",
                                "example": 4445
                            },
                            "total_stars": {
                                "type": "number",
                                "example": 4
                            }
                        }
                    }
                },
                "message": {
                    "type": "string",
                    "example": "Successfully retrieved recipes"
                },
                "status": {
                    "type": "string",
                    "example": "success"
                },
                "status_code": {
                    "type": "integer",
                    "example": 200
                },
                "success": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
...
        "models.GetRecipeByIdResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "type": "object",
                        "properties": {
                            "comments": {
                                "type": "array",
                                "items": {
                                    "type": "object",
                                    "properties": {
                                        "comment": {
                                            "type": "string",
                                            "example": "This is a comment"
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "example": "2021-01-01T00:00:00Z"
                                        },
                                        "created_by": {
                                            "type": "string",
                                            "example": "@facu.potti"
                                        },
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        }
                                    }
                                }
                            },
                            "created_by": {
                                "type": "string",
                                "example": "@facu.potti"
                            },
                            "description": {
                                "type": "string",
                                "example": "A feast for the senses"
                            },
                            "estimated_time": {
                                "type": "string",
                                "example": "20 minutes"
                            },
                            "id": {
                                "type": "integer",
                                "example": 1
                            },
                            "image_url": {
                                "type": "string",
                                "example": "https://example.com/image.png"
                            },
                            "ingredients": {
                                "type": "array",
                                "items": {
                                    "type": "object",
                                    "properties": {
                                        "description": {
                                            "type": "string",
                                            "example": "onion"
                                        },
                                        "name": {
                                            "type": "string",
                                            "example": "Chicken"
                                        },
                                        "quantity": {
                                            "type": "string",
                                            "example": "100"
                                        },
                                        "unit": {
                                            "type": "string",
                                            "example": "kg"
                                        }
                                    }
                                }
                            },
                            "instructions": {
                                "type": "array",
                                "items": {
                                    "type": "object",
                                    "properties": {
                                        "description": {
                                            "type": "string",
                                            "example": "Chop the chicken into small cubes or finely mince it (you can also use a food processor). This makes it easier to form patties later."
                                        },
                                        "image_url": {
                                            "type": "string",
                                            "example": "https://example.com/image.png"
                                        },
                                        "step": {
                                            "type": "string",
                                            "example": "1"
                                        }
                                    }
                                }
                            },
                            "title": {
                                "type": "string",
                                "example": "Chicken Wings"
                            },
                            "total_comments": {
                                "type": "integer",
                                "example": 64
                            },
                            "total_favorites": {
                                "type": "integer",
                                "example": 4445
                            },
                            "total_stars": {
                                "type": "number",
                                "example": 4
                            }
                        }
                    }
                },
                "message": {
                    "type": "string",
                    "example": "Successfully retrieved recipe"
                },
                "status": {
                    "type": "string",
                    "example": "success"
                },
                "status_code": {
                    "type": "integer",
                    "example": 200
                },
                "success": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
...
        "models.UpdateRecipeRequest": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string",
                    "example": "A feast for the senses"
                },
                "estimated_time": {
                    "type": "string",
                    "example": "20 minutes"
                },
                "image_url": {
                    "type": "string",
                    "example": "https://example.com/image.png"
                },
                "ingredients": {
                    "type": "array",
                    "items": {
                        "type": "object",
                        "properties": {
                            "description": {
                                "type": "string",
                                "example": "onion"
                            },
                            "name": {
                                "type": "string",
                                "example": "Chicken"
                            },
                            "quantity": {
                                "type": "string",
                                "example": "100"
                            },
                            "unit": {
                                "type": "string",
                                "example": "kg"
                            }
                        }
                    }
                },
                "instructions": {
                    "type": "array",
                    "items": {
                        "type": "object",
                        "properties": {
                            "description": {
                                "type": "string",
                                "example": "Chop the chicken into small cubes or finely mince it (you can also use a food processor). This makes it easier to form patties later."
                            },
                            "step": {
                                "type": "string",
                                "example": "1"
                            }
                        }
                    }
                },
                "title": {
                    "type": "string",
                    "example": "Chicken Wings"
                }
            }
        },
        "models.UpdateResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "Updated successfully"
                },
                "status": {
                    "type": "string",
                    "example": "success"
                },
                "status_code": {
                    "type": "integer",
                    "example": 200
                },
                "success": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
...
    }
}